requirements.txt
python-dotenv
pandas
boto3
python-dotenv